requirements.txt
asgiref==3.4.1
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
fastapi==0.67.0
greenlet==1.1.0
h11==0.12.0
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
PyYAML==5.4.1
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
